@@ -1,4 +1,11 @@
-part of whiteboard;
+import 'dart:typed_data';
+import 'dart:ui';
+
+import 'package:flutter/material.dart';
+import 'package:whiteboard/src/freehand_painter.dart';
+import 'package:whiteboard/src/models.dart';
+
+part 'whiteboard_controller.dart';
 
 typedef OnRedoUndo = void Function(bool isUndoAvailable, bool isRedoAvailable);
 
@@ -21,7 +28,6 @@ class WhiteBoard extends StatefulWidget {
 
   /// Callback for [Canvas] when it converted to image data.
   /// Use [WhiteBoardController] to convert.
-  final ValueChanged<Uint8List>? onConvertImage;
 
   /// This callback exposes if undo / redo is available and called successfully.
   final OnRedoUndo? onRedoUndo;
@@ -33,7 +39,6 @@ class WhiteBoard extends StatefulWidget {
     this.strokeColor = Colors.blue,
     this.strokeWidth = 4,
     this.isErasing = false,
-    this.onConvertImage,
     this.onRedoUndo,
   }) : super(key: key);
 
@@ -45,19 +50,19 @@ class _WhiteBoardState extends State<WhiteBoard> {
   final _undoHistory = <RedoUndoHistory>[];
   final _redoStack = <RedoUndoHistory>[];
 
-  final _strokes = <_Stroke>[];
+  final _strokes = <WhiteBoardStroke>[];
 
   // cached current canvas size
   late Size _canvasSize;
 
   // convert current canvas to image data.
-  Future<void> _convertToImage(ImageByteFormat format) async {
+  Future<Uint8List> _convertToImage(ImageByteFormat format) async {
     final recorder = PictureRecorder();
     final canvas = Canvas(recorder);
 
     // Emulate painting using _FreehandPainter
     // recorder will record this painting
-    _FreehandPainter(
+    FreehandPainter(
       _strokes,
       widget.backgroundColor,
     ).paint(canvas, _canvasSize);
@@ -71,53 +76,53 @@ class _WhiteBoardState extends State<WhiteBoard> {
     final converted =
         (await result.toByteData(format: format))!.buffer.asUint8List();
 
-    widget.onConvertImage?.call(converted);
+    return converted;
+  }
+
+  void onClear() {
+    if (_strokes.isEmpty) return;
+    setState(() {
+      final _removedStrokes = <WhiteBoardStroke>[]..addAll(_strokes);
+      final _redoUndoHistoryItem = RedoUndoHistory(
+        undo: () => setState(() => _strokes.addAll(_removedStrokes)),
+        redo: () => setState(() => _strokes.clear()),
+      );
+      _undoHistory.add(_redoUndoHistoryItem);
+      _strokes.clear();
+      _redoStack.clear();
+    });
+    widget.onRedoUndo?.call(_undoHistory.isNotEmpty, _redoStack.isNotEmpty);
+  }
+
+  bool onRedo() {
+    if (_redoStack.isEmpty) return false;
+
+    _undoHistory.add(_redoStack.removeLast()..redo());
+    widget.onRedoUndo?.call(_undoHistory.isNotEmpty, _redoStack.isNotEmpty);
+    return true;
+  }
+
+  bool onUndo() {
+    if (_undoHistory.isEmpty) return false;
+
+    _redoStack.add(_undoHistory.removeLast()..undo());
+    widget.onRedoUndo?.call(_undoHistory.isNotEmpty, _redoStack.isNotEmpty);
+    return true;
   }
 
   @override
   void initState() {
-    widget.controller?._delegate = _WhiteBoardControllerDelegate()
-      ..saveAsImage = _convertToImage
-      ..onUndo = () {
-        if (_undoHistory.isEmpty) return false;
-
-        _redoStack.add(_undoHistory.removeLast()..undo());
-        widget.onRedoUndo?.call(_undoHistory.isNotEmpty, _redoStack.isNotEmpty);
-        return true;
-      }
-      ..onRedo = () {
-        if (_redoStack.isEmpty) return false;
-
-        _undoHistory.add(_redoStack.removeLast()..redo());
-        widget.onRedoUndo?.call(_undoHistory.isNotEmpty, _redoStack.isNotEmpty);
-        return true;
-      }
-      ..onClear = () {
-        if (_strokes.isEmpty) return;
-        setState(() {
-          final _removedStrokes = <_Stroke>[]..addAll(_strokes);
-          _undoHistory.add(
-            RedoUndoHistory(
-              undo: () {
-                setState(() => _strokes.addAll(_removedStrokes));
-              },
-              redo: () {
-                setState(() => _strokes.clear());
-              },
-            ),
-          );
-          setState(() {
-            _strokes.clear();
-            _redoStack.clear();
-          });
-        });
-        widget.onRedoUndo?.call(_undoHistory.isNotEmpty, _redoStack.isNotEmpty);
-      };
+    widget.controller?._delegate = WhiteBoardControllerDelegate(
+      saveAsImage: _convertToImage,
+      onClear: onClear,
+      onUndo: onUndo,
+      onRedo: onRedo,
+    );
     super.initState();
   }
 
   void _start(double startX, double startY) {
-    final newStroke = _Stroke(
+    final newStroke = WhiteBoardStroke(
       color: widget.strokeColor,
       width: widget.strokeWidth,
       erase: widget.isErasing,
@@ -127,12 +132,8 @@ class _WhiteBoardState extends State<WhiteBoard> {
     _strokes.add(newStroke);
     _undoHistory.add(
       RedoUndoHistory(
-        undo: () {
-          setState(() => _strokes.remove(newStroke));
-        },
-        redo: () {
-          setState(() => _strokes.add(newStroke));
-        },
+        undo: () => setState(() => _strokes.remove(newStroke)),
+        redo: () => setState(() => _strokes.add(newStroke)),
       ),
     );
     _redoStack.clear();
@@ -147,7 +148,7 @@ class _WhiteBoardState extends State<WhiteBoard> {
 
   @override
   Widget build(BuildContext context) {
-    Size size = MediaQuery.of(context).size;
+    Size size = MediaQuery.sizeOf(context);
     return Container(
       height: size.height,
       width: size.width,
@@ -156,17 +157,15 @@ class _WhiteBoardState extends State<WhiteBoard> {
           details.localPosition.dx,
           details.localPosition.dy,
         ),
-        onPanUpdate: (details) {
-          _add(
-            details.localPosition.dx,
-            details.localPosition.dy,
-          );
-        },
+        onPanUpdate: (details) => _add(
+          details.localPosition.dx,
+          details.localPosition.dy,
+        ),
         child: LayoutBuilder(
           builder: (context, constraints) {
             _canvasSize = Size(constraints.maxWidth, constraints.maxHeight);
             return CustomPaint(
-              painter: _FreehandPainter(_strokes, widget.backgroundColor),
+              painter: FreehandPainter(_strokes, widget.backgroundColor),
             );
           },
         ),
