@@ -1,11 +1,3 @@
 library whiteboard;
 
-import 'dart:typed_data';
-import 'dart:ui';
-
-import 'package:flutter/material.dart';
-
-part 'src/whiteboard.dart';
-part 'src/WhiteBoardController.dart';
-part 'src/FreehandPainter.dart';
-part 'src/Model.dart';
+export 'src/whiteboard.dart' hide WhiteBoardControllerDelegate;
